Implement product detail features with ViewModel, repository caching, and UI
*   Change `Product` ID type from `Int` to `Long` across `ProductEntity`, `ProductDto`, and mappers.
*   Implement `ProductDetailsViewModel` to manage UI state and fetch product data using `SavedStateHandle` for navigation arguments.
*   Update `ProductRepository` and `ProductRepositoryImpl` to include `getProductDetails` with a caching strategy (10-minute timeout).
*   Expand `FakeStoreApi` with a new endpoint for fetching single product details.
*   Build out `ProductDetailScreen` using Jetpack Compose, featuring a split-view layout with an image header and detailed scrollable content.
*   Wire up `NavGraph` to use the new `ProductDetailsViewModel` for the detail screen.
*   Add unit tests in `ProductRepositoryImplTest` to verify cache-first and API fallback logic for product details.

+++ app/src/main/java/co/uk/pencepoint/data/local/entities/ProductEntity.kt
@@ -17,7 +17,7 @@ import androidx.room.PrimaryKey
  */
 @Entity(tableName = "products")
 data class ProductEntity(
-    @PrimaryKey(autoGenerate = false) val id: Int,
+    @PrimaryKey(autoGenerate = false) val id: Long,
     val title: String,
     val price: Double,
     val description: String,
